Move model data for settings to the Configuration namespace
