+++ PartyIcons/Configuration/ChatMode.cs
@@ -1,0 +1,8 @@
+﻿namespace PartyIcons.Configuration;
+
+public enum ChatMode
+{
+    GameDefault,
+    Role,
+    Job
+}
