+++ PartyIcons/Runtime/PartyListHUDUpdater.cs
@@ -17,7 +17,7 @@ public sealed class PartyListHUDUpdater : IDisposable
 {
     public bool UpdateHUD = false;
 
-    private readonly PluginConfiguration _configuration;
+    private readonly Settings _configuration;
     private readonly PartyListHUDView _view;
     private readonly RoleTracker _roleTracker;
 
@@ -30,7 +30,7 @@ public sealed class PartyListHUDUpdater : IDisposable
     private const string OpcodesUrl = "https://raw.githubusercontent.com/karashiiro/FFXIVOpcodes/master/opcodes.min.json";
     private List<int> _prepareZoningOpcodes = new();
 
-    public PartyListHUDUpdater(PartyListHUDView view, RoleTracker roleTracker, PluginConfiguration configuration)
+    public PartyListHUDUpdater(PartyListHUDView view, RoleTracker roleTracker, Settings configuration)
     {
         _view = view;
         _roleTracker = roleTracker;
