+++ PartyIcons/Runtime/ViewModeSetter.cs
@@ -20,13 +20,13 @@ public sealed class ViewModeSetter
     public bool InDuty { get; private set; } = true;
 
     private readonly NameplateView _nameplateView;
-    private readonly PluginConfiguration _configuration;
+    private readonly Settings _configuration;
     private readonly ChatNameUpdater _chatNameUpdater;
     private readonly PartyListHUDUpdater _partyListHudUpdater;
 
     private ExcelSheet<ContentFinderCondition> _contentFinderConditionsSheet;
 
-    public ViewModeSetter(NameplateView nameplateView, PluginConfiguration configuration, ChatNameUpdater chatNameUpdater,
+    public ViewModeSetter(NameplateView nameplateView, Settings configuration, ChatNameUpdater chatNameUpdater,
         PartyListHUDUpdater partyListHudUpdater)
     {
         _nameplateView = nameplateView;
