--- PartyIcons/View/ChatMode.cs
@@ -1,9 +1,0 @@
-﻿namespace PartyIcons.View;
-
-// public enum ChatMode
-// {
-//     GameDefault,
-//     OnlyColor,
-//     Role,
-//     Job
-// }
